PAPINO-36 При быстрой смене типа продукта может подгрузится из прошлого типа (так как после загрузки картинки происходит обновление переиспользуемой ячейки)
