+++ restaurant/src/main/java/ru/papino/restaurant/core/imageLoader/ImageLoaderImpl.kt
@@ -5,10 +5,9 @@ import android.graphics.drawable.Drawable
 import android.widget.ImageView
 import com.bumptech.glide.Glide
 import com.bumptech.glide.load.engine.DiskCacheStrategy
-import com.bumptech.glide.request.target.CustomTarget
-import com.bumptech.glide.request.transition.Transition
 
 class ImageLoaderImpl : ImageLoader {
+
     override fun loadImage(context: Context, view: ImageView, link: String) {
         Glide.with(context)
             .load(link)
@@ -21,21 +20,6 @@ class ImageLoaderImpl : ImageLoader {
             .asDrawable()
             .load(link)
             .diskCacheStrategy(DiskCacheStrategy.DATA)
-            .into(object : CustomTarget<Drawable>() {
-                override fun onResourceReady(
-                    resource: Drawable,
-                    transition: Transition<in Drawable>?
-                ) {
-                    onResource(resource)
-                }
-
-                override fun onLoadCleared(placeholder: Drawable?) {
-                    onResource(placeholder)
-                }
-
-                override fun onLoadFailed(errorDrawable: Drawable?) {
-                    onResource(errorDrawable)
-                }
-            })
+            .into(DrawableTarget(onResource))
     }
 }
